Functions for building IR.
The 'irbuilderref' type is no longer used. 'irnoderef' is the only newly
introduced Mu type. A bundle node now works as a namespace. All other
nodes are children of the bundle node. After a bundle node is loaded
into Mu (or aborted), Mu may de-allocate the bundle node and all of its
children for space efficiency.