Modified new_thread API and annotated muapi.h
The new_thread API call is split into two functions, one continues
normally and the other throws exception. This interface will be more
C-style.

It will be convenient if there is a program that parses the muapi.h and
generates language bindings (for Python, Scala, etc.). Some annotations
are added in the muapi.h file to make this process easier. For example,
when passing an array, C functions usually take two parameters: a
pointer and a size. The param:array:sz annotation tells the parser that
the "param" parameter points to an array whose size is determined by the
"sz" parameter. "param:bool" tells the parser that the "int param"
parameter holds a boolean value (non-zero or zero). "param:optional"
tells the parser that the "T* param" parameter may be NULL.